fix: enhanced otp handeling taxpayer apis
