+++ india_compliance/__init__.py
@@ -1,1 +1,3 @@
+from .exceptions import *
+
 __version__ = "16.0.0-dev"
