+++ india_compliance/public/js/utils.js
@@ -240,47 +240,6 @@ Object.assign(india_compliance, {
         }
     },
 
-    get_gstin_otp(company_gstin, error_type) {
-        let description = `An OTP has been sent to the registered mobile/email for GSTIN ${company_gstin} for further authentication. Please provide OTP.`;
-        if (error_type === "invalid_otp")
-            description = `Invalid OTP was provided for GSTIN ${company_gstin}. Please try again.`;
-
-        return new Promise(resolve => {
-            const prompt = new frappe.ui.Dialog({
-                title: __("Enter OTP"),
-                fields: [
-                    {
-                        fieldtype: "Data",
-                        label: __("One Time Password"),
-                        fieldname: "otp",
-                        reqd: 1,
-                        description: description,
-                    },
-                ],
-                primary_action_label: __("Submit"),
-                primary_action(values) {
-                    resolve(values.otp);
-                    prompt.hide();
-                },
-                secondary_action_label: __("Resend OTP"),
-                secondary_action() {
-                    frappe.call({
-                        method: "india_compliance.gst_india.utils.gstr_utils.request_otp",
-                        args: { company_gstin },
-                        callback: function () {
-                            frappe.show_alert({
-                                message: __("OTP has been resent."),
-                                indicator: "green",
-                            });
-                            prompt.get_secondary_btn().addClass("disabled");
-                        },
-                    });
-                },
-            });
-            prompt.show();
-        });
-    },
-
     guess_gst_category(gstin, country) {
         if (!gstin) {
             if (country && country !== "India") return "Overseas";
@@ -427,58 +386,6 @@ Object.assign(india_compliance, {
             .find(`.inner-group-button[data-label=${btn_group_name}]`)
             .find(`.dropdown-item[data-label="${encodeURIComponent(btn_name)}"]`)
             .addClass("text-danger");
-    },
-
-    async authenticate_company_gstins(company, company_gstin) {
-        const { message: gstin_authentication_status } = await frappe.call({
-            method: "india_compliance.gst_india.utils.gstr_utils.validate_company_gstins",
-            args: { company: company, company_gstin: company_gstin },
-        });
-
-        for (let gstin of Object.keys(gstin_authentication_status)) {
-            if (gstin_authentication_status[gstin]) continue;
-
-            gstin_authentication_status[gstin] =
-                await this.request_and_authenticate_otp(gstin);
-        }
-
-        return Object.keys(gstin_authentication_status);
-    },
-
-    async request_and_authenticate_otp(gstin) {
-        await frappe.call({
-            method: "india_compliance.gst_india.utils.gstr_utils.request_otp",
-            args: { company_gstin: gstin },
-        });
-
-        // wait for OTP to be authenticated to proceed
-        await this.authenticate_otp(gstin);
-    },
-
-    async authenticate_otp(gstin, error_type = null) {
-        if (!error_type) error_type = "otp_requested";
-
-        let is_authenticated = false;
-
-        while (!is_authenticated) {
-            const otp = await this.get_gstin_otp(gstin, error_type);
-
-            const { message } = await frappe.call({
-                method: "india_compliance.gst_india.utils.gstr_utils.authenticate_otp",
-                args: { company_gstin: gstin, otp: otp },
-            });
-
-            if (
-                message &&
-                ["otp_requested", "invalid_otp"].includes(message.error_type)
-            ) {
-                error_type = message.error_type;
-                continue;
-            }
-
-            is_authenticated = true;
-            return true;
-        }
     },
 
     show_dismissable_alert(wrapper, message, alert_type = "primary", on_close = null) {
